--- src/components/breadcrumb/Breadcrumb.utils.tsx
@@ -1,11 +1,0 @@
-import { BreadcrumbItem } from "./Breadcrumb.component";
-
-export const breadcrumbItems:Array<BreadcrumbItem> = [
-    {title: "Dashboard", path: "#"},
-    {title: "Profile", path: "/#/path"},
-    {title: "Details", path: "#",isActive:true}
-    ];
-
-export const breadcrumbItems1:Array<BreadcrumbItem> = [
-   {title: "Dashboard", path: "#",isActive:true},
-];
