+++ src/styles/components/Breadcrumb.scss
@@ -16,7 +16,6 @@
     display: block; 
     width: fit-content;
     padding: 23px 39px; 
-    background:$gray-100; 
     &__breadcrumb-item-active{
       @include breadcrumb-item($gray-900)  
    }
@@ -33,17 +32,19 @@
    list-style: none;
  }
  
- .breadcrumb-item + .breadcrumb-item {
-   padding-left: 0.5rem;
- }
-
- .breadcrumb-item + .breadcrumb-item::before {
-   float: left;
-   padding-right: 0.5rem;
-   color: $breadcrumb-separator-color;
-   content: var(--bs-breadcrumb-divider, "/");
- }
- 
- .breadcrumb-item.active {
-   color: $breadcrumb-separator-color;
- }
+.breadcrumb-item {
+	+ {
+		.breadcrumb-item {
+			padding-left: 0.5rem;
+			&::before {
+				float: left;
+				padding-right: 0.5rem;
+				color: $breadcrumb-separator-color;
+				content: var(--bs-breadcrumb-divider, "/");
+			}
+		}
+	}
+}
+.breadcrumb-item.active {
+	color: $breadcrumb-separator-color;
+}
